Factor out createDb, rename to createSdk
